// Shelf.WebApp/ShelfApp/src/app/shared/models/performance.ts
export interface Performance {
    IDPOG: number;
    IDPackage: number;
    IDProduct: number;
    Movt: number
    MovtAdj: number
    MovtLY: number
    NoOfStore: number
    PSW: number
    PerfPeriod: string
    Sales: number
    SalesLY: number
    TrueProfit: number
    UPC: string;
    SKU: string;
    Name: string;
    Package: string;
    HighlightLabel : string
}